refactor: 피드백 반영

+++ src/main/java/nextstep/qna/domain/Question.java
@@ -7,7 +7,8 @@ import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
-import static nextstep.qna.domain.ContentType.QUESTION;
+import static nextstep.qna.ExceptionMessage.DIFFERENT_WRITER_OF_QUESTION_AND_WRITER_OF_ANSWER;
+import static nextstep.qna.ExceptionMessage.NO_AUTHORITY_TO_DELETE_ANSWER;
 
 public class Question {
     private Long id;
@@ -62,15 +63,20 @@ public class Question {
     }
 
     public List<DeleteHistory> deleteByUser(NsUser user) throws CannotDeleteException {
-        validateUser(user);
+        validateDeletableQuestion(user);
         List<DeleteHistory> deleteHistoriesOfAnswers = answers.deleteByUser(user);
-        this.deleted = true;
+        updateDeletedAsTrue();
         return deleteHistoriesOfQuestion(deleteHistoriesOfAnswers);
+    }
+
+    private void validateDeletableQuestion(NsUser user) throws CannotDeleteException {
+        validateUser(user);
+        validateAnswers();
     }
 
     private void validateUser(NsUser user) throws CannotDeleteException {
         if (!isOwner(user)) {
-            throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
+            throw new CannotDeleteException(NO_AUTHORITY_TO_DELETE_ANSWER.message());
         }
     }
 
@@ -78,10 +84,20 @@ public class Question {
         return writer.equals(user);
     }
 
+    private void validateAnswers() throws CannotDeleteException {
+        if (!answers.isDeletableByWriter(writer)) {
+            throw new CannotDeleteException(DIFFERENT_WRITER_OF_QUESTION_AND_WRITER_OF_ANSWER.message());
+        }
+    }
+
+    private void updateDeletedAsTrue() {
+        this.deleted = true;
+    }
+
     private List<DeleteHistory> deleteHistoriesOfQuestion(List<DeleteHistory> deleteHistoriesOfAnswers) {
         List<DeleteHistory> deleteHistories = new ArrayList<>();
 
-        deleteHistories.add(new DeleteHistory(QUESTION, id, writer, LocalDateTime.now()));
+        deleteHistories.add(DeleteHistory.questionDeleteHistory(id, writer));
         deleteHistories.addAll(deleteHistoriesOfAnswers);
 
         return deleteHistories;
